exposed ID helper functions even more
